refactor : static LocalDate 빼기
