+++ Api/src/main/java/com/example/api/auth/service/helper/KakaoOauthHelper.java
@@ -5,12 +5,12 @@ import com.example.api.auth.model.dto.OauthUserInfoDto;
 import com.example.domains.user.enums.OauthInfo;
 import com.example.domains.user.enums.OauthProvider;
 import com.example.dto.OIDCDecodePayload;
-import com.example.oauth.client.KakaoInfoClient;
-import com.example.oauth.client.KakaoOauthClient;
-import com.example.oauth.dto.KakaoInfoResponseDto;
-import com.example.oauth.dto.KakaoTokenResponse;
-import com.example.oauth.dto.KakaoUnlinkTarget;
-import com.example.oauth.dto.OIDCPublicKeysResponse;
+import com.example.oauth.kakao.client.KakaoInfoClient;
+import com.example.oauth.kakao.client.KakaoOauthClient;
+import com.example.oauth.kakao.dto.KakaoInfoResponseDto;
+import com.example.oauth.kakao.dto.KakaoTokenResponse;
+import com.example.oauth.kakao.dto.KakaoUnlinkTarget;
+import com.example.oauth.kakao.dto.OIDCPublicKeysResponse;
 import com.example.properties.KakaoOAuthProperties;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
