+++ Infra/src/main/java/com/example/oauth/kakao/config/KakaoInfoConfig.java
@@ -1,4 +1,4 @@
-package com.example.oauth.config;
+package com.example.oauth.kakao.config;
 
 import feign.codec.Encoder;
 import feign.codec.ErrorDecoder;
@@ -6,13 +6,13 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
 
-@Import(KakaoKauthErrorDecoder.class)
-public class KakaoKauthConfig {
+@Import(KakaoInfoErrorDecoder.class)
+public class KakaoInfoConfig {
 
     @Bean
     @ConditionalOnMissingBean(value = ErrorDecoder.class)
-    public KakaoKauthErrorDecoder commonFeignErrorDecoder() {
-        return new KakaoKauthErrorDecoder();
+    public KakaoInfoErrorDecoder commonFeignErrorDecoder() {
+        return new KakaoInfoErrorDecoder();
     }
 
     @Bean
