+++ Infra/src/main/java/com/example/oauth/kakao/dto/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package com.example.oauth.dto;
+package com.example.oauth.kakao.dto;
 
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
