+++ Infra/src/main/java/com/example/oauth/kakao/dto/OIDCPublicKeyDto.java
@@ -1,4 +1,4 @@
-package com.example.oauth.dto;
+package com.example.oauth.kakao.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
